added user info to the UserDBConfig

+++ server/src/main/java/nl/tudelft/oopp/demo/services/BuildingService.java
@@ -28,7 +28,7 @@ public class BuildingService {
      * @param id we search for
      * @return the building with that exact id if it exists or null if not
      */
-    public Optional<Building> getBuilding(Integer id) {
+    public Optional<Building> getBuilding(Long id) {
         return buidlingRepository.findById(id);
     }
 
@@ -51,7 +51,7 @@ public class BuildingService {
      *
      * @param id to be deleted from the list of buildings
      */
-    public void delete(Integer id) {
+    public void delete(Long id) {
         buidlingRepository.deleteById(id);
     }
 }
